requirements.txt
anthropic>=0.50.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
